Move woocommerce tax flow to subdirectory
