+++ client/tasks/fills/tax/index.tsx
@@ -3,16 +3,9 @@
  */
 import { __ } from '@wordpress/i18n';
 import { Card, CardBody } from '@wordpress/components';
-import { difference } from 'lodash';
 import { useDispatch, useSelect } from '@wordpress/data';
-import { Spinner } from '@woocommerce/components';
 import { getAdminLink } from '@woocommerce/wc-admin-settings';
-import {
-	ONBOARDING_STORE_NAME,
-	OPTIONS_STORE_NAME,
-	PLUGINS_STORE_NAME,
-	SETTINGS_STORE_NAME,
-} from '@woocommerce/data';
+import { OPTIONS_STORE_NAME, SETTINGS_STORE_NAME } from '@woocommerce/data';
 import { queueRecordEvent } from '@woocommerce/tracks';
 import { registerPlugin } from '@wordpress/plugins';
 import { useEffect, useState } from '@wordpress/element';
@@ -21,17 +14,10 @@ import { WooOnboardingTask } from '@woocommerce/onboarding';
 /**
  * Internal dependencies
  */
-import {
-	AUTOMATION_PLUGINS,
-	hasCompleteAddress,
-	redirectToTaxSettings,
-	SettingsSelector,
-} from './utils';
-import { AutomatedTaxes } from './automated-taxes';
-import { ConfigurationStepper } from './configuration-stepper';
+import { redirectToTaxSettings, SettingsSelector } from './utils';
 import { createNoticesFromResponse } from '../../../lib/notices';
-import { getCountryCode } from '../../../dashboard/utils';
-import './tax.scss';
+import { PartnerCards } from './partner-cards';
+import { WooCommerceTax } from './woocommerce-tax/woocommerce-tax';
 
 const Tax = ( { onComplete, query } ) => {
 	const [ isPending, setIsPending ] = useState( false );
@@ -40,64 +26,17 @@ const Tax = ( { onComplete, query } ) => {
 	const { updateAndPersistSettingsForGroup } = useDispatch(
 		SETTINGS_STORE_NAME
 	);
-	const {
-		generalSettings,
-		isJetpackConnected,
-		isResolving,
-		pluginsToActivate,
-		tasksStatus,
-		taxSettings,
-	} = useSelect( ( select ) => {
+	const { generalSettings, taxSettings } = useSelect( ( select ) => {
 		const { getSettings } = select(
 			SETTINGS_STORE_NAME
 		) as SettingsSelector;
-		const { getActivePlugins } = select( PLUGINS_STORE_NAME );
-		const activePlugins = getActivePlugins();
 
 		return {
 			generalSettings: getSettings( 'general' ).general,
-			isJetpackConnected: select(
-				PLUGINS_STORE_NAME
-			).isJetpackConnected(),
-			isResolving:
-				! select( PLUGINS_STORE_NAME ).hasFinishedResolution(
-					'isJetpackConnected'
-				) ||
-				! select(
-					SETTINGS_STORE_NAME
-				).hasFinishedResolution( 'getSettings', [ 'general' ] ) ||
-				! select( ONBOARDING_STORE_NAME ).hasFinishedResolution(
-					'getTasksStatus'
-				),
-			pluginsToActivate: difference( AUTOMATION_PLUGINS, activePlugins ),
 			// @Todo this should be removed as soon as https://github.com/woocommerce/woocommerce-admin/pull/7841 is merged.
-			tasksStatus: select( ONBOARDING_STORE_NAME ).getTasksStatus(),
 			taxSettings: getSettings( 'tax' ).tax || {},
 		};
 	} );
-
-	const supportsAutomatedTaxes = () => {
-		const {
-			automatedTaxSupportedCountries = [],
-			taxJarActivated,
-		} = tasksStatus;
-
-		return (
-			! taxJarActivated && // WCS integration doesn't work with the official TaxJar plugin.
-			automatedTaxSupportedCountries.includes(
-				getCountryCode( generalSettings?.woocommerce_default_country )
-			)
-		);
-	};
-
-	const canAutomateTaxes = () => {
-		return (
-			hasCompleteAddress( generalSettings ) &&
-			! pluginsToActivate.length &&
-			isJetpackConnected &&
-			supportsAutomatedTaxes()
-		);
-	};
 
 	const onManual = async () => {
 		setIsPending( true );
@@ -172,26 +111,21 @@ const Tax = ( { onComplete, query } ) => {
 		}
 	}, [] );
 
-	if ( isResolving ) {
-		return <Spinner />;
-	}
-
 	const childProps = {
 		isPending,
 		onAutomate,
 		onManual,
 		onDisable,
-		supportsAutomatedTaxes: supportsAutomatedTaxes(),
 	};
 
 	return (
 		<div className="woocommerce-task-tax">
 			<Card className="woocommerce-task-card">
 				<CardBody>
-					{ canAutomateTaxes() ? (
-						<AutomatedTaxes { ...childProps } />
+					{ query.partner === 'woocommerce-tax' ? (
+						<WooCommerceTax { ...childProps } />
 					) : (
-						<ConfigurationStepper { ...childProps } />
+						<PartnerCards { ...childProps } />
 					) }
 				</CardBody>
 			</Card>
