+++ client/tasks/fills/tax/partner-cards.tsx
@@ -1,0 +1,17 @@
+/**
+ * External dependencies
+ */
+import { __ } from '@wordpress/i18n';
+
+/**
+ * Internal dependencies
+ */
+
+export const PartnerCards = ( {
+	isPending,
+	onAutomate,
+	onManual,
+	onDisable,
+} ) => {
+	return <div>Partner cards</div>;
+};
